refactor(kaspi): extract status routes
